requirements.txt
# Reddit Persona Generator - Fixed Dependencies
# This version avoids protobuf compatibility issues

# Core dependencies
torch>=1.13.0
transformers>=4.24.0,<4.35.0
tokenizers>=0.13.0,<0.15.0

# Protobuf - specific version to avoid MessageFactory issues
protobuf>=3.20.0,<3.21.0

# Quantization support
bitsandbytes>=0.35.0
accelerate>=0.15.0

# Reddit API
praw>=7.6.0

# NLP and text processing
spacy>=3.4.0
nltk>=3.7
textblob>=0.17.1

# Web framework
flask>=2.2.0
flask-cors>=3.0.10

# Utilities
python-dotenv>=0.19.0
numpy>=1.21.0
scipy>=1.9.0

# Optional: For better GPU support
# nvidia-ml-py3>=7.352.0

# Note: Install spaCy English model separately:
# python -m spacy download en_core_web_sm 